Update doc /filtering code for duplicate mutations
Thanks @jingchunzhu for the relevant documentation on the Xena Browser Google
Group. See https://groups.google.com/d/msg/ucsc-cancer-genomics-browser/eg6nJOFSefw/Z0BM6pU9BAAJ:

> regarding the duplicated rows, they are due to that there are multiple
sequencing results for these samples, from the  multiple TCGA aliquots map to a
single sample. We left the duplications in the file intentionally, if you see
the same mutation is called on two rows, it is essentially a biological
replicate.  And you should also see occasionally in these duplications, majority
of them has, say for example, two exact same rows for sample A, but sometime
there is only one row of mutation B for sample A, which means that mutation B is
not detected in the biological replicate.